updated images to more recent examples, see changelog
