+++ .gitignore
@@ -4,6 +4,7 @@ __pycache__/
 *$py.class
 
 deploy/
+useragent.ini
 
 # C extensions
 *.so
