+++ emotes/all_emotes.txt
