+++ emotes/emojis.txt
